+++ CakeMail.RestClient/Models/ClientRegistrationInfo.cs
@@ -7,19 +7,19 @@ namespace CakeMail.RestClient.Models
 		#region Properties
 
 		[JsonProperty("client_id")]
-		public int ClientId { get; set; }
+		public long ClientId { get; set; }
 
 		[JsonProperty("client_key")]
 		public string ClientKey { get; set; }
 
 		[JsonProperty("admin_id")]
-		public int AdminId { get; set; }
+		public long AdminId { get; set; }
 
 		[JsonProperty("admin_key")]
 		public string AdminKey { get; set; }
 
 		[JsonProperty("contact_id")]
-		public int PrimaryContactId { get; set; }
+		public long PrimaryContactId { get; set; }
 
 		[JsonProperty("contact_key")]
 		public string PrimaryContactKey { get; set; }
