+++ CakeMail.RestClient/Models/Template.cs
@@ -5,21 +5,24 @@ using System.Collections.Generic;
 
 namespace CakeMail.RestClient.Models
 {
+	/// <summary>
+	/// A Template is HTML or Text based content that may be stored for later use.
+	/// </summary>
 	public class Template
 	{
 		#region Properties
 
 		[JsonProperty("id")]
-		public int Id { get; set; }
+		public long Id { get; set; }
 
 		[JsonProperty("labels")]
 		public KeyValuePair<string, string>[] Labels { get; set; }
-		
+
 		[JsonProperty("status")]
 		public string Status { get; set; }
 
 		[JsonProperty("category_id")]
-		public int CategoryId { get; set; }
+		public long CategoryId { get; set; }
 
 		[JsonProperty("type")]
 		public string Type { get; set; }
@@ -35,7 +38,7 @@ namespace CakeMail.RestClient.Models
 		public DateTime ModifiedOn { get; set; }
 
 		[JsonProperty("editor_version")]
-		public int EditorVersion { get; set; }
+		public long EditorVersion { get; set; }
 
 		[JsonProperty("name")]
 		public string Name { get; set; }
